FIxes, formatting, unlock sound

+++ hammerspoon/modules/kill_helpers_on_quit.lua
@@ -1,5 +1,6 @@
 local module = {}
-local appProcessPairs, appWatcher
+local appWatcher
+local appProcessPairs
 
 local function killProcess(processToKill)
   hs.execute("pkill -9 '" .. processToKill .. "'")
@@ -7,9 +8,7 @@ end
 
 function module.killAllNow()
   if appProcessPairs and #appProcessPairs > 0 then
-    for _, pair in ipairs(appProcessPairs) do
-      killProcess(pair.processToKill)
-    end
+    for _, pair in ipairs(appProcessPairs) do killProcess(pair.processToKill) end
   end
 end
 
@@ -23,9 +22,7 @@ function module.init(config)
       if eventType == hs.application.watcher.terminated then
         local appName = appObject:name()
         for _, appProcessPair in ipairs(appProcessPairs) do
-          if appName == appProcessPair.appName then
-            killProcess(appProcessPair.processToKill)
-          end
+          if appName == appProcessPair.appName then killProcess(appProcessPair.processToKill) end
         end
       end
     end):start()
